+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Instructions follows [Semantic Versioning](http://semver.org/).
 ## 1.1.0
 Released on 2017-08-XX.
 
+### Added
+- Add a new delegate method to handle tap on the overlay ([#100]).
+
 ### Fixed
 - Fix [#127] & [#132] by retrieve configuration from parent controller.